rhinometric-console/collector/requirements.txt
requests>=2.28
psutil>=5.9
pyyaml>=6.0
